// src/API.ts
export enum AFFILIATIONSTATUS {
  DRAFT = "DRAFT",
  PENDING = "PENDING",
  ACTIVE = "ACTIVE",
  ARCHIVED = "ARCHIVED",
}
// ...
export enum ROOMTYPE {
  KING1 = "KING1",
  QUEEN2 = "QUEEN2",
  DOUBLE2 = "DOUBLE2",
  OTHER = "OTHER",
}
// ...
export enum SERVICEMEMBERSTATUS {
  ACTIVEDUTY = "ACTIVEDUTY",
  RESERVE = "RESERVE",
  NATIONALGUARD = "NATIONALGUARD",
  VETERAN = "VETERAN",
}
// ...
export enum GUESTTYPE {
  PRIMARY = "PRIMARY",
  ADDITIONAL = "ADDITIONAL",
}